+++ src/Data/Map/Monoidal/Strict.hs
@@ -3,8 +3,10 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE InstanceSigs #-}
 {-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 
@@ -132,6 +134,10 @@ import Prelude hiding (null, lookup, map, foldl, foldr, filter)
 import Data.Coerce (coerce)
 import Data.Semigroup
 import Data.Foldable (Foldable)
+import Data.Functor.Apply (Apply)
+import Data.Functor.Alt (Alt(..))
+import Data.Functor.Bind (Bind)
+import Data.Functor.Plus (Plus)
 import Data.Traversable (Traversable)
 import Control.Applicative (Applicative, pure)
 import Data.Data (Data)
@@ -162,6 +168,7 @@ newtype MonoidalMap k a = MonoidalMap { getMonoidalMap :: M.Map k a }
     deriving ( Show, Read, Functor, Eq, Ord, NFData
              , Foldable, Traversable
              , FromJSON, ToJSON, FromJSON1, ToJSON1
+             , Apply, Bind, Plus
              , Data, Typeable, Align
 #if MIN_VERSION_these(0,8,0)
              , Semialign
@@ -172,6 +179,7 @@ newtype MonoidalMap k a = MonoidalMap { getMonoidalMap :: M.Map k a }
 #endif
 #endif
              )
+
 
 #if MIN_VERSION_containers(0,5,9)
 deriving instance (Ord k) => Eq1 (MonoidalMap k)
@@ -229,6 +237,10 @@ instance (Ord k, Semigroup a) => Monoid (MonoidalMap k a) where
     mappend (MonoidalMap a) (MonoidalMap b) = MonoidalMap $ M.unionWith (<>) a b
     {-# INLINE mappend #-}
 #endif
+
+instance Ord k => Alt (MonoidalMap k) where
+  (<!>) :: forall k a. Ord k => MonoidalMap k a -> MonoidalMap k a -> MonoidalMap k a
+  (<!>) = coerce ((<!>) :: M.Map k a -> M.Map k a -> M.Map k a)
 
 instance Newtype (MonoidalMap k a) (M.Map k a) where
     pack = MonoidalMap
@@ -697,5 +709,3 @@ maxViewWithKey = coerce (M.maxViewWithKey :: M.Map k a -> Maybe ((k, a), M.Map k
 valid :: forall k a. Ord k => MonoidalMap k a -> Bool
 valid = coerce (M.valid :: Ord k => M.Map k a -> Bool)
 {-# INLINE valid #-}
-
-